\name{grades}
\alias{grades}
\title{grades data}
\usage{
  data(grades)
}
\description{
  Prints 'grades data'. The scores are the all student of the Biology from Biostatistical courses, from Universidad Nacional, Costa Rica.
}

\format{An object call \code{grades} with 47 rows and 2 columns.}

\value{
  \item{scores}{is numeric (scores)}

  \item{year}{is factor (year of the courses)}
}


\examples{
  data(grades)
}


\keyword{datasets}
