\name{resumeR}
\alias{resumeR}
\title{Resume for R}

\description{
  Resume for R, is a package for R, that perform a complete statistical summary. The idea is to facilitate the researcher to obtain fast and agile information.
}
\author{
  Oscar Ramírez-Alán. Universidad Nacional, Escuela de Ciencias Biológicas.
}
