\name{scores}
\alias{scores}
\title{scores data}
\usage{
  data(scores)
}
\description{
  Prints 'scores data'. The scores are the all student of the Biology from Biostatistical courses, from Universidad Nacional, Costa Rica.
  }


\arguments{
  \item{scores}{is numeric (scores)}

  \item{year}{is factor (year of the courses)}
}


\examples{
  data(scores)
}
